manual-ingester/pyproject.toml
[project]
name = "manual-ingester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "claude-agent-sdk>=0.1.48",
    "python-dotenv>=1.2.2",
    "unsiloed-sdk>=0.1.7",
    "voyageai>=0.3.7",
]
